+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 - Add ability to set lighthouses to standby
-- Fix connections not being cleaned up
+- Improve error handling and connection management
 - Fix window size issues happening on some systems
 
 ## Version 0.1.0 (2025-07-06)